server/requirements.txt
alembic==1.16.4
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.5
numpy==2.3.2
openai==1.98.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
wordcloud==1.9.4
